Split alternative runners into modules
